correcting addListItem
Naming consistency: Ensure that the function addListItem is properly named and used throughout the code.

Incorrect syntax for adding addListItem to the returned object: The syntax addListItem(pokemon): addListItem is incorrect. You should use a comma instead of a colon to separate the properties of the returned object.

Missing function call in the return statement: You need to include the addListItem function in the object being returned from the IIFE.

Incorrect usage of addListItem inside forEach loop: You should call the addListItem function inside the forEach loop, passing the pokemon object as an argument.